+++ main.go
@@ -9,6 +9,7 @@ import (
 	"syscall"
 
 	"git.sr.ht/~jrswab/akordo/controller"
+	"git.sr.ht/~jrswab/akordo/plugins"
 	"git.sr.ht/~jrswab/akordo/roles"
 	"git.sr.ht/~jrswab/akordo/xp"
 	"github.com/bwmarrin/discordgo"
@@ -37,7 +38,13 @@ func loadSavedData(sd *controller.SessionData) {
 	// Load saved self assign role data into the struct created by NewSessionData
 	if _, err := os.Stat(roles.SelfAssignFile); err == nil {
 		if err := sd.Roles.LoadSelfAssignRoles(roles.SelfAssignFile); err != nil {
-			log.Fatalf("error loading role file: %s", err)
+			log.Fatalf("error loading self assign role file: %s", err)
+		}
+	}
+	// Load saved banned word data into the struct created by NewSessionData
+	if _, err := os.Stat(plugins.BannedWordsPath); err == nil {
+		if err := sd.Blacklist.LoadBannedWordList(plugins.BannedWordsPath); err != nil {
+			log.Fatalf("error loading banned words file: %s", err)
 		}
 	}
 }