+++ CHANGELOG.md
@@ -8,9 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Clear command to clear messages of the bot or a user for the past 100 messages.
-- Deletion of send command message after the bot returns the request
+- Deletion of command message after the bot returns the request
+- The ability to add banned words.
+- bannedWords.json to .gitignore
+- Removal of user from the guild when a blacklisted word is used.
 
 ### Changed
+- Controller.go `msgType` to be "chan" by default.
+- `checkSyntax()` to `checkMessage()`
+- the controller reply method to be unexported
 
 ### Fixed
 - Bug where xp was looking at the bot token instead of the bot ID environment variable.